+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 - fixed crash related to Explosive Arrows shot from dispensers
 - fixed crash on Mohist servers
 - added Rarity Property
+- fixed compat with Connector
+- added Apoli compat property